pyproject.toml
[tool.poetry]
name = "repl_python3_Rage-Bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
audioread = "^2.1.9"
"discord.py" = { version = "^1.0.1",  extras = ["voice"]}
flask = "^2.0.1"
replit = "^3.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
